refactor(interpreter): move interpreter_common.rs to common/
