+++ src/query/service/src/interpreters/interpreter_privilege_revoke.rs
@@ -19,7 +19,7 @@ use common_meta_types::PrincipalIdentity;
 use common_planner::plans::RevokePrivilegePlan;
 use common_users::UserApiProvider;
 
-use crate::interpreters::interpreter_common::validate_grant_object_exists;
+use crate::interpreters::common::validate_grant_object_exists;
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
 use crate::sessions::QueryContext;
