+++ src/query/service/src/interpreters/interpreter_user_stage_remove.rs
@@ -20,7 +20,7 @@ use common_exception::Result;
 use common_planner::plans::RemoveStagePlan;
 use regex::Regex;
 
-use crate::interpreters::interpreter_common::list_files;
+use crate::interpreters::common::list_files;
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
 use crate::sessions::QueryContext;
